+++ rvc/lib/algorithm/discriminators/sub/__init__.py
@@ -1,0 +1,4 @@
+from .period import DiscriminatorP  # noqa
+from .scale import DiscriminatorS  # noqa
+from .band import DiscriminatorB  # noqa
+from .cqt import DiscriminatorCQT  # noqa
